artificial dataset does nt«ot delete

+++ src/artificial_dataset/main_artificial_dataset.py
@@ -12,37 +12,47 @@ sys.path.insert(0, 'src/common')
 from Variables import *
 from Util import *
 
-# manage folders
-create_folders(path_main_dataset)
-create_folders(path_to_images_folder)
-create_folders(path_to_statistics_gt_folder)
-create_folders(path_to_statistics_pred_folder)
-create_folders(path_to_outputs_folder)
-create_folders(path_to_masks_overlapped_pred_folder)
-create_folders(path_to_masks_single_pred_folder)
-create_folders(path_to_masks_overlapped_gt_folder)
-create_folders(path_to_masks_single_gt_folder)
+# Prompt the user for their name
+deleteDataset = input("Do you wanna delete the old dataset? (y)es or (n)o: ")
+if deleteDataset == "y":
+    deleteDataset = input("Sure? (y)es or (n)o: ")
+    
+    # manage folders
+    create_folders(path_main_dataset)
+    create_folders(path_to_images_folder)
+    create_folders(path_to_statistics_gt_folder)
+    create_folders(path_to_statistics_pred_folder)
+    create_folders(path_to_outputs_folder)
+    create_folders(path_to_masks_overlapped_pred_folder)
+    create_folders(path_to_masks_single_pred_folder)
+    create_folders(path_to_masks_overlapped_gt_folder)
+    create_folders(path_to_masks_single_gt_folder)
 
 
-delete_folder_contents(path_to_images_folder)
-delete_folder_contents(path_to_statistics_gt_folder)
-delete_folder_contents(path_to_statistics_pred_folder)
-delete_folder_contents(path_to_outputs_folder)
-delete_folder_contents(path_to_masks_overlapped_pred_folder)
-delete_folder_contents(path_to_masks_single_pred_folder)
-delete_folder_contents(path_to_masks_overlapped_gt_folder)
-delete_folder_contents(path_to_masks_single_gt_folder)
+    delete_folder_contents(path_to_images_folder)
+    delete_folder_contents(path_to_statistics_gt_folder)
+    delete_folder_contents(path_to_statistics_pred_folder)
+    delete_folder_contents(path_to_outputs_folder)
+    delete_folder_contents(path_to_masks_overlapped_pred_folder)
+    delete_folder_contents(path_to_masks_single_pred_folder)
+    delete_folder_contents(path_to_masks_overlapped_gt_folder)
+    delete_folder_contents(path_to_masks_single_gt_folder)
 
+    index = 0
+
+if deleteDataset == "n":
+    index = input("What is the last index? ")
+    index = int(index) + 1
 
 colors = Colors()
-today_date = str(datetime.now().date())
 
 # generate images
 for i in range(num_wsp):
     start_time = time.time()
+    filename = i + index
 
-    print("Creating image number ", i)
-    wsp_image = WSP_Image(i, colors, today_date)
+    print("Creating image number ", i + index)
+    wsp_image = WSP_Image(filename, colors)
     print("Image created. Calculating statistics...")
     WSP_Statistics(wsp_image, colors)
 
@@ -52,4 +62,3 @@ for i in range(num_wsp):
     print("Time taken:", elapsed_time, "seconds")
 
     
-## Background:  188.33333333333334
